[tool] Warn and continue if unable to add key or key title to diagram
A packaging bug reported in #147 revealed that if and when the renderer
was unable to add the key or key title to the diagram, it would totally
fail. I’d prefer graceful degradation wherein it’d still produce a
diagram, just without the key (or the key title) if there’s a problem
with those specific elements.