[tool] Move edit CLI workflow into the main codebase
So that I can test it, because I found a bug, but before I fix it I need
to be able to test it.

(By convention the code in the source tree src/cli is not tested. I
think this is maybe mostly OK, if we are rigorous about keeping the
amount of code in it very small. In this case, there was already too
much code in fc4.cli.edit, so this is a good move; total lines of code
in src/cli dropped from 226 to 164.)