[tool] Reproduce concurrency bug in edit workflow
The bug is that if two or more files are changed more-or-less
simultaneously, and rendering then takes more than a second or two
(which it generally does) then the workflow will enter an infinite loop
wherein it changes the files and those changes are then picked up and
trigger another change, et cetera, ad nauseum.

There’s some logic in fc4.io.edit that’s supposed to guard against this,
but it’s not working in this case — that’s the bug.

I’ll explain the cause of the bug in the next commit, wherein I’ll fix it.