[tool] Improve fix to concurrency bug in edit workflow
While working on the tests I realized that because the ns has ~mutable
state in vars, the ns is essentially a context for *shared mutable
state* 😱 and that this would probably cause problems if the tests were
run in parallel, which I’d like them to be.

So I refactored the ns such that all the state is stored in the watch
map — there is no longer any state in vars. The only real API change is
that `stop` now requires that it be passed the watch, and this is fine.