[tool] Improve the testability of the edit workflow
I found that when I ran each test individually in my REPL, they passed,
but when I used eftest to run them concurrently, they failed. I realized
that this was due to them both printing to stdout and thus the output of
each test could not be examined individually.

So I found a way to ensure that they’d use different writers, one for
each test.