[docs] Add “…and documentarians…” to target audience
I’m going to demo the framework in an hour here at Write the Docs
Portland 2019 and I want everyone here to feel included in the target
audience!

This is a little bit of an ambitious expansion of the target audience,
but I feel good about it. Ambitious in a good way; I think it’s (mostly)
ready.